Cast arguments to avoid SunPro warnings.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@1805 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/src/xatoi.c
@@ -26,7 +26,7 @@ int xatoi(int *i, const char *in) {
 	errno = ERANGE;
 	return 0;
     }
-    *i = val;
+    *i = (int)val; /* safe after range check */
     return 1;
 }
